Add meta plugin to prevent needing peer dependencies.
This creates a `metalab` plugin that coalesces all the rules from all the other plugins and allows configuring of those rules through the `metalab/` prefix. e.g. `import/no-commonjs` becomes `metalab/import/no-commonjs`. This prevents users from having to install 10 different eslint plugins as dependencies and avoids all the `peerDependency` warnings that can happen when versions get bumped.

In addition to copying `rules` we might want to copy `deprecatedRules` too.

Note that since `eslint` forces you to not be able to resolve plugins locally and forces the `eslint-plugin-` prefix we have to do a little tomfoolery and copy `plugin` into the `node_modules` folder ourselves on install. It's not ideal but it appears to work.

Handful of rule changes here since this is a major version bump. Nothing huge but of note:

 * JSX pragma back to `React`,
 * Filenames must match exports,
 * Use babelized versions of rules when appropriate,
 * Relax function styles,
 * Remove need for sorting component methods,
 * Add `jsx-a11y`,
 * Add `flowtype`.